Serial hamlib fix

+++ src/setupdialog.cpp
@@ -35,7 +35,7 @@ This class calls all the othet "Setup..." to manage the configuration
 
 SetupDialog::SetupDialog(DataProxy_SQLite *dp, const QString &_configFile, const QString &_softwareVersion, const int _page, const bool _firstTime, QWidget *parent)
 {
-    //qDebug() << "SetupDialog::SetupDialog 2: " << _configFile << "/" << _softwareVersion << "/" << QString::number(_page) << util->boolToQString(_firstTime) << endl ;
+    //qDebug() << Q_FUNC_INFO << ": " << _configFile << "/" << _softwareVersion << "/" << QString::number(_page) << util->boolToQString(_firstTime) << endl ;
 
     logSeverity = Info;
     constrid = 2;
@@ -47,40 +47,36 @@ SetupDialog::SetupDialog(DataProxy_SQLite *dp, const QString &_configFile, const
     version = _softwareVersion;
     pageRequested = _page;
     int logsPageTabN=-1;
-    //qDebug() << "SetupDialog::SetupDialog 01" << QT_ENDL;
+    //qDebug() << Q_FUNC_INFO << ": 01" << QT_ENDL;
 
     locator = new Locator();
 
     tabWidget = new QTabWidget;
-    //qDebug() << "SetupDialog::SetupDialog 01.0" << QT_ENDL;
+    //qDebug() << Q_FUNC_INFO << ": 01.0" << QT_ENDL;
     userDataPage = new SetupPageUserDataPage(dataProxy);
-    //qDebug() << "SetupDialog::SetupDialog 01.10" << QT_ENDL;
+    //qDebug() << Q_FUNC_INFO << ": 01.10" << QT_ENDL;
     bandModePage = new SetupPageBandMode(dataProxy, this);
-    //qDebug() << "SetupDialog::SetupDialog 01.20" << QT_ENDL;
+    //qDebug() << Q_FUNC_INFO << ": 01.20" << QT_ENDL;
     dxClusterPage = new SetupPageDxCluster(this);
-    //qDebug() << "SetupDialog::SetupDialog 01.30" << QT_ENDL;
+    //qDebug() << Q_FUNC_INFO << ": 01.30" << QT_ENDL;
     colorsPage = new SetupPageColors(this);
-    //qDebug() << "SetupDialog::SetupDialog 01.40" << QT_ENDL;
+    //qDebug() << Q_FUNC_INFO << ": 01.40" << QT_ENDL;
     miscPage = new SetupPageMisc(this);
-    //qDebug() << "SetupDialog::SetupDialog 01.50" << QT_ENDL;
+    //qDebug() << Q_FUNC_INFO << ": 01.50" << QT_ENDL;
     worldEditorPage = new SetupPageWorldEditor (dataProxy, this);
-    //qDebug() << "SetupDialog::SetupDialog 01.60" << QT_ENDL;
+    //qDebug() << Q_FUNC_INFO << ": 01.60" << QT_ENDL;
     logsPage = new SetupPageLogs(dataProxy, this);
-    //qDebug() << "SetupDialog::SetupDialog 01.70" << QT_ENDL;
+    //qDebug() << Q_FUNC_INFO << ": 01.70" << QT_ENDL;
     eLogPage = new SetupPageELog(this);
-    //qDebug() << "SetupDialog::SetupDialog 01.80" << QT_ENDL;
+    //qDebug() << Q_FUNC_INFO << ": 01.80" << QT_ENDL;
     UDPPage = new SetupPageUDP(this);
-    //qDebug() << "SetupDialog::SetupDialog 01.90" << QT_ENDL;
+    //qDebug() << Q_FUNC_INFO << ": 01.90" << QT_ENDL;
     satsPage = new SetupPageSats(dataProxy, this);
-    //qDebug() << "SetupDialog::SetupDialog 01.100" << QT_ENDL;
+    //qDebug() << Q_FUNC_INFO << ": 01.100" << QT_ENDL;
     hamlibPage = new SetupPageHamLib(dataProxy, this);
     logViewPage = new SetupPageLogView(dataProxy, this);
-    //subdivisionsPage = new SetupPageSubdivisions(dataProxy, this);
-    //regionalAwardsPage = new SetupPageRegionalAwards(dataProxy, this);
-    //interfacesWindowsPage = new SetupPageInterfacesWindows(this);
+    //qDebug() << Q_FUNC_INFO << ": 02" << QT_ENDL;
 
-    //qDebug() << "SetupDialog::SetupDialog 02" << QT_ENDL;
-    //tabWidget->addTab(subdivisionsPage, tr("Subdivisions"));
     tabWidget->addTab(userDataPage, tr("User data"));
     tabWidget->addTab(bandModePage, tr("Bands/Modes"));
     tabWidget->addTab(logViewPage, tr("Log widget"));
@@ -90,14 +86,10 @@ SetupDialog::SetupDialog(DataProxy_SQLite *dp, const QString &_configFile, const
     tabWidget->addTab(worldEditorPage, tr("World Editor"));
     logsPageTabN = tabWidget->addTab(logsPage, tr("Logs"));
     tabWidget->addTab(eLogPage, tr("eLog"));
-    //tabWidget->addTab(clubLogPage, tr("ClubLog"));
-    //tabWidget->addTab(lotwPage, tr("LoTW"));
-
     tabWidget->addTab(UDPPage, tr("WSJT-X"));
     tabWidget->addTab(satsPage , tr("Satellites"));
+    //qDebug() << Q_FUNC_INFO << ": 02.100" << QT_ENDL;
     tabWidget->addTab(hamlibPage, tr ("HamLib"));
-    //tabWidget->addTab(interfacesWindowsPage, tr ("Interfaces"));
-
     //qDebug() << "SetupDialog::SetupDialog 03" << QT_ENDL;
 
     QPushButton *closeButton = new QPushButton(tr("Cancel"));
@@ -115,33 +107,33 @@ SetupDialog::SetupDialog(DataProxy_SQLite *dp, const QString &_configFile, const
     mainLayout->addLayout(horizontalLayout);
     mainLayout->addLayout(buttonsLayout);
 
-    //qDebug() << "SetupDialog::SetupDialog 04" << QT_ENDL;
+    //qDebug() << Q_FUNC_INFO << ": 04" << QT_ENDL;
 
     setLayout(mainLayout);
     setWindowTitle(tr("Settings"));
 
-    //qDebug() << "SetupDialog::SetupDialog 05" << QT_ENDL;
+    //qDebug() << Q_FUNC_INFO << ": 05" << QT_ENDL;
 
     slotReadConfigData();
-    //qDebug() << "SetupDialog::SetupDialog 05.1" << QT_ENDL;
+    //qDebug() << Q_FUNC_INFO << ": 05.1" << QT_ENDL;
 
     if ((pageRequested==6) && (logsPageTabN>0))// The user is opening a new log
     {
-         //qDebug() << "SetupDialog::SetupDialog 05.2" << QT_ENDL;
+        //qDebug() << Q_FUNC_INFO << ": 5.2" << QT_ENDL;
         tabWidget->setCurrentIndex(logsPageTabN);
     }
-    //qDebug() << "SetupDialog::SetupDialog 05.3" << QT_ENDL;
+    //qDebug() << Q_FUNC_INFO << ": 5.3" << QT_ENDL;
     nolog = !(haveAtleastOneLog());
 
     connect(closeButton, SIGNAL(clicked()), this, SLOT(slotCancelButtonClicked()));
     connect(okButton, SIGNAL(clicked()), this, SLOT(slotOkButtonClicked()));
     connectActions();
-    //qDebug() << "SetupDialog::SetupDialog 2  - END" << QT_ENDL;
+    //qDebug() << Q_FUNC_INFO << " - END" << QT_ENDL;
 }
 
 SetupDialog::~SetupDialog()
 {
-       //qDebug() << "SetupDialog::~SetupDialog " << QT_ENDL;
+    //qDebug() << Q_FUNC_INFO  << QT_ENDL;
 }
 
 void SetupDialog::connectActions()
@@ -1002,15 +994,15 @@ bool SetupDialog::processConfigLine(const QString &_line)
         hamlibPage->setSerialPort(value);
     }else if(tab =="HAMLIBSERIALBAUDS"){
           //qDebug() << "SetupDialog::processConfigLine: HAMLIBSERIALBAUDS: " << value << QT_ENDL;
-        hamlibPage->setSerialSpeed(value);
+        hamlibPage->setSerialSpeed(value.toInt());
     }else if(tab =="HAMLIB"){
           //qDebug() << "SetupDialog::processConfigLine: HAMLIB: " << value << QT_ENDL;
         hamlibPage->setActive(value);
     }else if(tab=="HAMLIBREADONLY"){
         hamlibPage->setReadOnly(value);
     }else if(tab =="HAMLIBSERIALDATABITS"){
-          //qDebug() << "SetupDialog::processConfigLine: HAMLIBSERIALDATABITS: " << value << QT_ENDL;
-        hamlibPage->setDataBits(value);
+        //qDebug() << "SetupDialog::processConfigLine: HAMLIBSERIALDATABITS: " << value << QT_ENDL;
+        hamlibPage->setDataBits(value.toInt ());
     }else if(tab =="HAMLIBSERIALSTOPBITS"){
           //qDebug() << "SetupDialog::processConfigLine: HAMLIBSERIALSTOPBITS: " << value << QT_ENDL;
         hamlibPage->setStopBits(value);
@@ -1174,6 +1166,7 @@ void SetupDialog::readActiveBands (const QString &actives)
     bands << dataProxy->getBandsInLog(-1);
     bands << _abands;
     bands.removeDuplicates();
+    //qDebug() << Q_FUNC_INFO << " - END";
     emit debugLog (Q_FUNC_INFO, "END", logSeverity);
 }
